[win]#136 Adds Missing TMX and Spriter dependencies. Fixes compiler warnings treated as errors in a hacky way. Base64 conversion is not to be trusted on x64 machines...

+++ IndieLib/common/include/IND_SpriterEntity.h
@@ -112,7 +112,7 @@ private:
     
     int                         _currentAnimation;      // current animation playing
     int                         _currentKey;            // current key of animation playing
-    int                         _currentTime;           // current time of the animation
+    double                         _currentTime;           // current time of the animation
     
     bool                        _drawBones;             // TODO: support this in a later version
     bool                        _drawObjectpositions;   // TODO: support this in a later version
